requirements.txt
unidecode
rank_bm25
streamlit
google-cloud-firestore
uuid
datetime
spacy
firebase_admin
pyrebase